+++ include/platform/esp32/ess_platform_esp32.h
@@ -39,16 +39,20 @@
 #include "platform/ess_platform_interface.h"
 
 
-class ess_platform_esp32 : public ess_platform_interface<ess_platform_esp32>  {
-  friend class ess_platform_interface<ess_platform_esp32> ;
+class ess_platform_esp32 : public ess_platform_interface<ess_platform_esp32, ESS_CONFIC_MAX_CORES>  {
+  friend class ess_platform_interface<ess_platform_esp32, ESS_CONFIC_MAX_CORES> ;
+  friend void _esp32_platform_montoring_cpu0( void * parameter ) ;
+  friend void _esp32_platform_montoring_cpu1( void * parameter ) ;
 protected:
   ess_platform_esp32();
 public:
   static ess_format_t get_format() { return ESS_DEFAULT_SERVER_FORMAT; }
 
+  virtual ess_error_t create() ;
 
   virtual ess_output_module* create_output(ess_output_type type, std::string controller_name,
     ess_format_t format = ESS_DEFAULT_SERVER_FORMAT);
+
 };
 /**
 * @}