+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 
 ## Release History
+* 0.8.10:
+  - add ess_error_t ess_task::start(uint32_t core)
+  - add ess_audio_block_t
+  - update read functions in channel and modules virtual unsigned int  read(ess_audioblock_t *block, unsigned int offset)
+  - add memory map for ess_audio_block_t
+  - add new configs:
+    - ESS_MAX_AUDIO_MEMORY 163840,
+    - ESS_MEMORY_MAP_DEBUG
+    - ESS_PLATFORM_MONTORING 1 /**< when set then useble get_cpu_max and get_cpu_load*/
 * 0.8.01:
   - add  class ess_stereo_simple_buffer_output_module
     mix floated audio data to a simple integer buffer stereo